// worklenz-frontend/src/app/administrator/reporting/interfaces.ts
export interface IRPTDuration {
  label: string;
  key: string;
  dates?: string
}

export interface IReportingInfo {
  organization_name: string;
}

export interface IRPTTeamStatistics {
  count: number;
  projects: number;
  members: number;
}

export interface IRPTProjectStatistics {
  count: number;
  active: number;
  overdue: number;
}

export interface IRPTMemberStatistics {
  count: number;
  unassigned: number;
  overdue: number;
}

export interface ITimeLogBreakdownReq {
  id: string,
  date_range: string[],
  duration: string,
  time_zone: string
}
// ...
export interface IProjectLogs {
  project_name: string;
  task_key: string;
  task_name: string;
  time_spent_string: string;
  user_name: string;
  avatar_url: string;
  created_at: string;
}

export interface IProjectLogsBreakdown {
  log_day: string;
  logs: IProjectLogs[]
}
// ...
export interface IRPTTimeProject {
  id: string;
  name: string;
  color_code: string;
  value?: number;
  estimated_value?: number;
  end_date?: string;
}